+++ src/components/ColorMainComponent.tsx
@@ -1,17 +1,19 @@
 import { motion } from 'framer-motion';
-import { CopyToClipboard } from 'react-copy-to-clipboard';
-import { HiOutlineColorSwatch } from 'react-icons/hi';
-import { IoIosCopy } from 'react-icons/io';
+import { IoIosBrush, IoIosCopy } from 'react-icons/io';
 import tinycolor from 'tinycolor2';
 
-import notify from '@/lib/toast';
+import CopyButton from './CopyButton';
 function ColorMainComponent({ color, type, onClick }) {
   const colorObj = tinycolor(color);
+  const rgb = tinycolor(color).toRgb();
   return (
     <motion.div
       layoutId={`picker-${type}`}
-      className='flex flex-row-reverse overflow-hidden rounded-xl shadow-2xl shadow-stone-400/20 sm:flex-col lg:flex-row-reverse xl:flex-col'
-      style={{ backgroundColor: colorObj.toHexString() }}
+      className='flex flex-row-reverse overflow-hidden rounded-xl sm:flex-col lg:flex-row-reverse xl:flex-col'
+      style={{
+        backgroundColor: colorObj.toHexString(),
+        boxShadow: `0 25px 50px -12px rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, 0.25), 0 10px 50px -12px rgba(0, 0, 0, 0.08)`,
+      }}
     >
       <motion.button
         layoutId={`picker-area-${type}`}
@@ -20,7 +22,7 @@ function ColorMainComponent({ color, type, onClick }) {
         aria-label={`Click to edit ${type} color`}
         title={`Click to edit ${type} color`}
       >
-        <HiOutlineColorSwatch
+        <IoIosBrush
           className='h-6 w-6 opacity-0 transition-opacity group-hover:opacity-100 md:h-8 md:w-8'
           style={{ color: colorObj.isDark() ? '#FFF' : '#000' }}
         />
@@ -29,17 +31,7 @@ function ColorMainComponent({ color, type, onClick }) {
         <p className='text-xs font-semibold capitalize leading-none text-stone-800 dark:text-stone-100'>
           {type}
         </p>
-        <CopyToClipboard
-          text={colorObj.toHexString()}
-          onCopy={() =>
-            notify({
-              style: {
-                backgroundColor: colorObj.toHexString(),
-                color: colorObj.isDark() ? '#FFF' : '#000',
-              },
-            })
-          }
-        >
+        <CopyButton color={colorObj.toHexString()}>
           <button
             className='group flex w-full items-center justify-between pt-1 font-mono text-xs text-stone-500 transition-colors hover:text-primary-500 active:text-primary-800 dark:text-stone-400 dark:active:text-primary-200'
             aria-label={`Click to copy ${colorObj.toHexString()}`}
@@ -48,7 +40,7 @@ function ColorMainComponent({ color, type, onClick }) {
             {colorObj.toHexString()}
             <IoIosCopy className='h-3 w-3 opacity-0 transition-opacity group-hover:opacity-100' />
           </button>
-        </CopyToClipboard>
+        </CopyButton>
       </div>
     </motion.div>
   );