+++ src/components/ColorPicker.tsx
@@ -34,7 +34,7 @@ function ColorPicker({ type, colors, onChange, onDismiss }) {
     >
       <motion.div
         layoutId={`picker-${type}`}
-        className='flex flex-col gap-4 overflow-hidden rounded-xl bg-white text-left shadow-2xl dark:bg-stone-900'
+        className='relative flex flex-col gap-4 rounded-xl bg-white text-left shadow-2xl shadow-stone-900/20 dark:bg-stone-900'
       >
         <h4 className='z-[2] px-4 pt-4 capitalize text-stone-800 dark:text-stone-100'>
           {type} Color
@@ -47,18 +47,26 @@ function ColorPicker({ type, colors, onChange, onDismiss }) {
           />
         </motion.div>
         <motion.div className='px-4'>
-          <input
-            type='text'
-            name='color'
-            id='color'
-            className='block w-full rounded-full border-gray-300 bg-transparent px-4 text-stone-800 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 dark:text-stone-100 sm:text-sm'
-            placeholder='Enter a color'
-            value={currentColor}
-            onChange={handleInputChange}
-          />
+          <form
+            onSubmit={(e) => {
+              e.preventDefault();
+              handleInputChange({ target: { value: e.target[0].value } });
+            }}
+          >
+            <input
+              type='text'
+              name='color'
+              id='color'
+              className='block w-full rounded-full border-gray-300 bg-transparent px-4 text-stone-800 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 dark:text-stone-100 sm:text-sm'
+              placeholder='Enter a color'
+              autoFocus
+              value={currentColor}
+              onChange={handleInputChange}
+            />
+          </form>
         </motion.div>
         <button
-          className='block w-full border-t border-t-stone-400/20 px-4 py-2 text-stone-800 transition hover:bg-stone-400/20 dark:text-stone-100'
+          className='block w-full rounded-b-xl border-t border-t-stone-400/20 px-4 py-2 text-stone-800 transition hover:bg-stone-400/20 dark:text-stone-100'
           onClick={onDismiss}
         >
           Done