+++ src/components/ColorPillComponent.tsx
@@ -1,8 +1,7 @@
-import { CopyToClipboard } from 'react-copy-to-clipboard';
 import { IoIosCopy } from 'react-icons/io';
 import tinycolor from 'tinycolor2';
 
-import notify from '@/lib/toast';
+import CopyButton from './CopyButton';
 
 function ColorPillComponent({ color, type }) {
   const monos =
@@ -18,17 +17,7 @@ function ColorPillComponent({ color, type }) {
       {monos.map((mono, i) => {
         return (
           <div key={`${type}.${i}.${mono.toRgbString()}`} className='relative'>
-            <CopyToClipboard
-              text={mono.toHexString()}
-              onCopy={() =>
-                notify({
-                  style: {
-                    backgroundColor: mono.toHexString(),
-                    color: mono.isDark() ? '#FFF' : '#000',
-                  },
-                })
-              }
-            >
+            <CopyButton color={mono.toRgbString()}>
               <button
                 className='group flex h-10 w-full items-center justify-between rounded-full px-4 dark:border dark:border-white/10'
                 style={{ backgroundColor: mono.toHexString() }}
@@ -36,7 +25,7 @@ function ColorPillComponent({ color, type }) {
                 title={`Click to copy ${mono.toHexString()}`}
               >
                 <div
-                  className='flex w-full flex-wrap justify-between px-0.5 text-left text-xs opacity-20 transition-opacity group-hover:opacity-100 pointer-touch:opacity-90'
+                  className='flex w-full flex-wrap justify-between px-0.5 text-left text-xs opacity-30 transition-opacity group-hover:opacity-100 pointer-touch:opacity-90'
                   style={{ color: mono.isDark() ? '#fff' : '#000' }}
                 >
                   {/* <div className='font-mono lowercase'>{mono.toRgbString()}</div> */}
@@ -47,7 +36,7 @@ function ColorPillComponent({ color, type }) {
                   </div>
                 </div>
               </button>
-            </CopyToClipboard>
+            </CopyButton>
           </div>
         );
       })}