+++ CHANGELOG.md
@@ -1,21 +1,55 @@
+
 # Changelog
+
+  
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [1.0.1](https://github.com/fluid-design-io/color-ui-generator/compare/v0.1.1...v1.0.1) (2022-04-24)
+  
 
-### 0.1.1 (2022-04-24)
-
-## 1.0.0 - 2022-04-24
+## 1.1.0 - 2020-04-24
 
 ### New Features
 
-- #### Inital Release
-    
-    Inital 1.0.0 release
+ - Added support to copy colors
+ - Added Github page link
+ - Updated favicon
+  
 
 ### Improvements & Bug Fixes
 
-- #### Fix: Color UI Generator
-    
-    Fixed a bug where the color UI generator was not working
+- Fixed some components not having a dark theme
+- Improved general UI
+
+  
+  
+
+## [1.0.1](https://github.com/fluid-design-io/color-ui-generator/compare/v0.1.1...v1.0.1) (2022-04-24)
+
+  
+
+## 0.1.1 (2022-04-24)
+
+  
+
+## 1.0.0 - 2022-04-24
+
+  
+
+### New Features
+
+  
+
+-  #### Inital Release
+
+Inital 1.0.0 release
+
+  
+
+### Improvements & Bug Fixes
+
+  
+
+-  #### Fix: Color UI Generator
+
+Fixed a bug where the color UI generator was not working